requirements.txt
alembic==1.14.0
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
flake8==7.1.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Pydantic==0.12.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-dotenv==0.5.2
pytest-mock==3.14.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
